+++ aeon/forecasting/deep_learning/_deepar.py
@@ -18,7 +18,7 @@ from sklearn.utils import check_random_state
 
 from aeon.forecasting.base import DirectForecastingMixin, IterativeForecastingMixin
 from aeon.forecasting.deep_learning.base import BaseDeepForecaster
-from aeon.networks._deepar import DeepARNetwork
+from aeon.networks.auto_encoder._deepar import DeepARNetwork
 
 
 class DeepARForecaster(
