+++ aeon/networks/__init__.py
@@ -22,22 +22,22 @@ __all__ = [
     "DeepARNetwork",
     "TCNNetwork",
 ]
-from aeon.networks._ae_abgru import AEAttentionBiGRUNetwork
-from aeon.networks._ae_bgru import AEBiGRUNetwork
-from aeon.networks._ae_dcnn import AEDCNNNetwork
-from aeon.networks._ae_drnn import AEDRNNNetwork
-from aeon.networks._ae_fcn import AEFCNNetwork
-from aeon.networks._ae_resnet import AEResNetNetwork
-from aeon.networks._cnn import TimeCNNNetwork
-from aeon.networks._dcnn import DCNNNetwork
-from aeon.networks._deepar import DeepARNetwork
-from aeon.networks._disjoint_cnn import DisjointCNNNetwork
-from aeon.networks._encoder import EncoderNetwork
-from aeon.networks._fcn import FCNNetwork
-from aeon.networks._inception import InceptionNetwork
-from aeon.networks._lite import LITENetwork
-from aeon.networks._mlp import MLPNetwork
-from aeon.networks._resnet import ResNetNetwork
-from aeon.networks._rnn import RecurrentNetwork
-from aeon.networks._tcn import TCNNetwork
+from aeon.networks.auto_encoder._ae_abgru import AEAttentionBiGRUNetwork
+from aeon.networks.auto_encoder._ae_bgru import AEBiGRUNetwork
+from aeon.networks.auto_encoder._ae_dcnn import AEDCNNNetwork
+from aeon.networks.auto_encoder._ae_drnn import AEDRNNNetwork
+from aeon.networks.auto_encoder._ae_fcn import AEFCNNetwork
+from aeon.networks.auto_encoder._ae_resnet import AEResNetNetwork
 from aeon.networks.base import BaseDeepLearningNetwork
+from aeon.networks.encoder._cnn import TimeCNNNetwork
+from aeon.networks.encoder._dcnn import DCNNNetwork
+from aeon.networks.encoder._disjoint_cnn import DisjointCNNNetwork
+from aeon.networks.encoder._encoder import EncoderNetwork
+from aeon.networks.encoder._fcn import FCNNetwork
+from aeon.networks.encoder._inception import InceptionNetwork
+from aeon.networks.encoder._lite import LITENetwork
+from aeon.networks.encoder._mlp import MLPNetwork
+from aeon.networks.encoder._resnet import ResNetNetwork
+from aeon.networks.encoder._rnn import RecurrentNetwork
+from aeon.networks.encoder._tcn import TCNNetwork
+from aeon.networks.encoder_decoder._deepar import DeepARNetwork
