+++ aeon/networks/auto_encoder/_ae_bgru.py
@@ -25,8 +25,7 @@ class AEBiGRUNetwork(BaseDeepLearningNetwork):
     """
 
     _config = {
-        "python_dependencies": ["tensorflow"],
-        "python_version": "<3.13",
+        **BaseDeepLearningNetwork._config,
         "structure": "auto-encoder",
     }
