+++ aeon/networks/auto_encoder/_ae_dcnn.py
@@ -53,8 +53,7 @@ class AEDCNNNetwork(BaseDeepLearningNetwork):
     """
 
     _config = {
-        "python_dependencies": ["tensorflow"],
-        "python_version": "<3.13",
+        **BaseDeepLearningNetwork._config,
         "structure": "auto-encoder",
     }
