+++ aeon/networks/auto_encoder/_ae_drnn.py
@@ -57,8 +57,7 @@ class AEDRNNNetwork(BaseDeepLearningNetwork):
     """
 
     _config = {
-        "python_dependencies": ["tensorflow"],
-        "python_version": "<3.13",
+        **BaseDeepLearningNetwork._config,
         "structure": "auto-encoder",
     }
