+++ aeon/networks/auto_encoder/_ae_resnet.py
@@ -69,8 +69,7 @@ class AEResNetNetwork(BaseDeepLearningNetwork):
     """
 
     _config = {
-        "python_dependencies": ["tensorflow"],
-        "python_version": "<3.13",
+        **BaseDeepLearningNetwork._config,
         "structure": "auto-encoder",
     }
