+++ aeon/networks/auto_encoder/tests/test_ae_fcn.py
