+++ aeon/networks/encoder/_cnn.py
@@ -54,8 +54,7 @@ class TimeCNNNetwork(BaseDeepLearningNetwork):
     """
 
     _config = {
-        "python_dependencies": ["tensorflow"],
-        "python_version": "<3.13",
+        **BaseDeepLearningNetwork._config,
         "structure": "encoder",
     }
