+++ aeon/networks/encoder/_disjoint_cnn.py
@@ -85,8 +85,7 @@ class DisjointCNNNetwork(BaseDeepLearningNetwork):
     """
 
     _config = {
-        "python_dependencies": ["tensorflow"],
-        "python_version": "<3.13",
+        **BaseDeepLearningNetwork._config,
         "structure": "encoder",
     }
