+++ aeon/networks/encoder/_encoder.py
@@ -43,6 +43,11 @@ class EncoderNetwork(BaseDeepLearningNetwork):
     for Artificial Intelligence, 120--129 2018.
     """
 
+    _config = {
+        **BaseDeepLearningNetwork._config,
+        "structure": "encoder",
+    }
+
     def __init__(
         self,
         kernel_size=None,
