+++ aeon/networks/encoder/_fcn.py
@@ -1,24 +1,19 @@
-"""Auto-Encoder using Fully Convolutional Network (FCN)."""
+"""Fully Convolutional Network (FCNNetwork)."""
 
 __maintainer__ = ["hadifawaz1999"]
 
-import numpy as np
 
 from aeon.networks.base import BaseDeepLearningNetwork
 
 
-class AEFCNNetwork(BaseDeepLearningNetwork):
-    """Establish the network structure for a AE-FCN.
+class FCNNetwork(BaseDeepLearningNetwork):
+    """
+    Establish the network structure for a FCN.
 
-    Auto-Encoder based Fully Convolutional Netwwork (AE-FCN),
-    adapted from the implementation used in [1]_.
+    Adapted from the implementation used in [1]_
 
     Parameters
     ----------
-    latent_space_dim : int, default = 128
-        Dimension of the auto-encoder's latent space.
-    temporal_latent_space : bool, default = False
-        Flag to choose whether the latent space is an MTS or Euclidean space.
     n_layers : int, default = 3
         Number of convolution layers.
     n_filters : int or list of int, default = [128,256,128]
@@ -43,29 +38,17 @@ class AEFCNNetwork(BaseDeepLearningNetwork):
 
     References
     ----------
-    .. [1] Network originally defined in:
-    @inproceedings{wang2017time,
-      title={Time series classification from scratch with deep neural networks:
-       A strong baseline},
-      author={Wang, Zhiguang and Yan, Weizhong and Oates, Tim},
-      booktitle={2017 International joint conference on neural networks
-      (IJCNN)},
-      pages={1578--1585},
-      year={2017},
-      organization={IEEE}
-    }
+    .. [1] Wang et al. Time series classification from scratch with deep neural
+    networks: a strong baseline, IJCNN, 2017
     """
 
     _config = {
-        "python_dependencies": ["tensorflow"],
-        "python_version": "<3.13",
-        "structure": "auto-encoder",
+        **BaseDeepLearningNetwork._config,
+        "structure": "encoder",
     }
 
     def __init__(
         self,
-        latent_space_dim=128,
-        temporal_latent_space=False,
         n_layers=3,
         n_filters=None,
         kernel_size=None,
@@ -75,8 +58,6 @@ class AEFCNNetwork(BaseDeepLearningNetwork):
         activation="relu",
         use_bias=True,
     ):
-        self.latent_space_dim = latent_space_dim
-        self.temporal_latent_space = temporal_latent_space
         self.n_layers = n_layers
         self.n_filters = n_filters
         self.kernel_size = kernel_size
@@ -93,13 +74,14 @@ class AEFCNNetwork(BaseDeepLearningNetwork):
 
         Parameters
         ----------
-        input_shape : tuple of shape = (n_timepoints (m), n_channels (d))
-            The shape of the data fed into the input layer.
+        input_shape : tuple
+          shape = (n_timepoints (m), n_channels (d)), the shape of the data fed
+          into the input layer.
 
         Returns
         -------
-        encoder : a keras Model.
-        decoder : a keras Model.
+        input_layer : a keras layer
+        output_layer : a keras layer
         """
         import tensorflow as tf
 
@@ -124,7 +106,6 @@ class AEFCNNetwork(BaseDeepLearningNetwork):
                     f" the same as number of layers but is"
                     f" not: {self.n_layers}"
                 )
-            self._kernel_size = self.kernel_size
             self._kernel_size = self._kernel_size_
         else:
             self._kernel_size = [self._kernel_size_] * self.n_layers
@@ -184,9 +165,9 @@ class AEFCNNetwork(BaseDeepLearningNetwork):
         else:
             self._use_bias = [self.use_bias] * self.n_layers
 
-        input_layer_encoder = tf.keras.layers.Input(input_shape)
+        input_layer = tf.keras.layers.Input(input_shape)
 
-        x = input_layer_encoder
+        x = input_layer
 
         for i in range(self.n_layers):
             conv = tf.keras.layers.Conv1D(
@@ -199,78 +180,10 @@ class AEFCNNetwork(BaseDeepLearningNetwork):
             )(x)
 
             conv = tf.keras.layers.BatchNormalization()(conv)
-            conv = tf.keras.layers.Activation(
-                activation=self._activation[i], name=f"__act_encoder_block{i}"
-            )(conv)
+            conv = tf.keras.layers.Activation(activation=self._activation[i])(conv)
 
             x = conv
 
-        if not self.temporal_latent_space:
-            shape_before_flattent = x.shape[1:]
-            flatten_layer = tf.keras.layers.Flatten()(x)
-            latent_space = tf.keras.layers.Dense(units=self.latent_space_dim)(
-                flatten_layer
-            )
-        else:
-            latent_space = tf.keras.layers.Conv1D(
-                filters=self.latent_space_dim,
-                kernel_size=1,
-                strides=self._strides[-1],
-                padding=self._padding[-1],
-                dilation_rate=self._dilation_rate[-1],
-                use_bias=self._use_bias[-1],
-            )(x)
+        gap_layer = tf.keras.layers.GlobalAveragePooling1D()(conv)
 
-        encoder = tf.keras.models.Model(
-            inputs=input_layer_encoder, outputs=latent_space, name="encoder"
-        )
-
-        if not self.temporal_latent_space:
-            input_layer_decoder = tf.keras.layers.Input((self.latent_space_dim,))
-
-            # Cast to int to avoid Keras rejecting numpy scalar types
-            decoder_units = int(np.prod(shape_before_flattent))
-            dense_layer = tf.keras.layers.Dense(units=decoder_units)(
-                input_layer_decoder
-            )
-
-            reshape_layer = tf.keras.layers.Reshape(target_shape=shape_before_flattent)(
-                dense_layer
-            )
-            x = reshape_layer
-        else:
-            input_layer_decoder = tf.keras.layers.Input(latent_space.shape[1:])
-
-            x = input_layer_decoder
-
-        for i in range(self.n_layers)[::-1]:
-            conv = tf.keras.layers.Conv1DTranspose(
-                filters=self._n_filters[i],
-                kernel_size=self._kernel_size[i],
-                strides=self._strides[i],
-                dilation_rate=self._dilation_rate[i],
-                padding=self._padding[i],
-                use_bias=self._use_bias[i],
-            )(x)
-
-            conv = tf.keras.layers.BatchNormalization()(conv)
-            conv = tf.keras.layers.Activation(
-                activation=self._activation[i], name=f"__act_decoder_block{i}"
-            )(conv)
-
-            x = conv
-
-        last_projection_layer = tf.keras.layers.Conv1DTranspose(
-            filters=input_shape[-1],
-            kernel_size=1,
-            padding=self._padding[0],
-            strides=self._strides[0],
-            dilation_rate=self._dilation_rate[0],
-            use_bias=self._use_bias[0],
-        )(x)
-
-        decoder = tf.keras.models.Model(
-            inputs=input_layer_decoder, outputs=last_projection_layer, name="decoder"
-        )
-
-        return encoder, decoder
+        return input_layer, gap_layer
