+++ aeon/networks/encoder/_resnet.py
@@ -63,8 +63,7 @@ class ResNetNetwork(BaseDeepLearningNetwork):
     """
 
     _config = {
-        "python_dependencies": ["tensorflow"],
-        "python_version": "<3.13",
+        **BaseDeepLearningNetwork._config,
         "structure": "encoder",
     }
