+++ aeon/networks/encoder/tests/test_dcnn.py
