+++ aeon/networks/encoder/tests/test_disjoint_cnn.py
