+++ aeon/networks/encoder/tests/test_fcn.py
