+++ aeon/networks/encoder/tests/test_tcn.py
