+++ aeon/networks/encoder/tests/test_time_cnn.py
